# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Devices Service for technical challenge.",
        "title": "Devices Service",
        "contact": {
            "name": "Victor Springer"
        },
        "license": {
            "name": "MIT License"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/devices": {
            "get": {
                "description": "Get a list of all devices",
                "produces": [
                    "application/json"
                ],
                "summary": "List all devices",
                "operationId": "list-all-devices",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Creates a new device",
                "produces": [
                    "application/json"
                ],
                "summary": "Add device",
                "operationId": "add-device",
                "parameters": [
                    {
                        "description": "Device to add",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/device.Device"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/devices/search": {
            "get": {
                "description": "Get a list of device data by brand",
                "produces": [
                    "application/json"
                ],
                "summary": "Get devices by brand",
                "operationId": "search-devices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device's brand",
                        "name": "brand",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/devices/{id}": {
            "get": {
                "description": "Get device data by id",
                "produces": [
                    "application/json"
                ],
                "summary": "Get device by id",
                "operationId": "get-device-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device's ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "Delete device data by id",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete device",
                "operationId": "delete-device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device's ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "patch": {
                "description": "Update device data by id",
                "produces": [
                    "application/json"
                ],
                "summary": "Update device",
                "operationId": "update-device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device's ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "device",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/device.Device"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "device.Device": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "creationTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updateTime": {
                    "type": "string"
                }
            }
        }
    }
}